Refactor model classes
