+++ src/com/cicerone/model/QuestionType.java
@@ -1,0 +1,5 @@
+package com.cicerone.model;
+
+public enum QuestionType {
+    SINGLE_CHOICE, MULTIPLE_CHOICE, TRUE_OR_FALSE
+}
