--- src/model/QuestionType.java
@@ -1,5 +1,0 @@
-package model;
-
-public enum QuestionType {
-    DESCRIPTIVE, MULTIPLE_CHOICE, TRUE_OR_FALSE
-}
